requirements.txt
setuptools==33.1.1
Django==1.11
djangorestframework==3.6.2
django-constance==2.0.0
django-picklefield==0.3.2
dj-database-url==0.4.2
gunicorn==19.6.0
ipython==5.2.2
Pillow==3.4.2
PyMySQL==0.7.11
Unipath==1.1

# Debug
django-querycount==0.5.0